refactor: FSD구조변경 [DF-94]
